refactor(forge): split into lib + thin binary, fix what that exposed
Same crate-shape problem as cmd/waf. zhenai-forge was a binary crate with no
lib.rs, so `pub` granted no dead-code exemption and 84 of its 96 clippy
warnings were dead_code on test-exercised research paths, cascading false
positives, and ADR-048's deliberate ForgeBackend extension surface. Adding
src/lib.rs and reducing main.rs to CLI dispatch takes the crate 96 -> 0.

Cargo infers the lib name from the package name, so no [lib] stanza is needed.
The split cost three compile errors, all the same shape: main.rs reached into
gemma4_gpu's pub(crate) profiling helpers, which are now genuinely library API
since the CLI is an external consumer.

Real findings behind the noise:

- gemma4.rs computed the layer-0 embedding twice on every backward pass. A
  redundant `let layer_input = ...` called compute_embed_input and threw the
  result away through `let _ =`, under the author's own "clean up" note.
  NOTE: test_gemma4_backward_grad_health, which would exercise this end to
  end, is #[ignore]d because it needs the deleted Gemma-4 GGUF. The numerical
  backward tests pass and the rewrite is a direct equivalence, but it is not
  covered by a running test.

- GpuTrainer::new and forward_loss were pub while taking &CpuWeights, whose
  fields and loader are both private — no caller outside the crate could ever
  construct one. Narrowed the methods to pub(crate) rather than widening the
  type to paper over the leak.

Smaller: three no-op *mut c_void casts in hip.rs, a missing is_empty beside
GpuBuffer::len, checked_div for two hand-written divide guards, clamp for a
.max(50).min(200), and type aliases NormedInputs / LazyAttnCache /
AttnInputGrads for signatures flagged as very complex.

Four items are genuinely unused and kept deliberately, each annotated with its
reason rather than a blanket allow: CpuWeights::forward_loss and
n_total_layers belong to the Mistral path CLAUDE.md retains under the "side
stuff" rule; per_head_rmsnorm_cpu and gelu_mul_batch_on_gpu are a CPU
validation reference and a working-but-unwired fused-GELU wrapper.

Reasoning record in notes/dead-code-is-a-crate-shape-problem.md. 104 tests
pass.

+++ crates/zhenai-forge/src/lora.rs
@@ -96,8 +96,10 @@ impl LoraLayer {
 
         // Step 1: hidden = A × input (rank,)
         let mut hidden = vec![0.0f32; self.rank as usize];
+        #[allow(clippy::needless_range_loop)] // strided tensor index — see crate note
         for r in 0..self.rank as usize {
             let mut sum = 0.0f32;
+            #[allow(clippy::needless_range_loop)] // strided tensor index — see crate note
             for d in 0..self.input_dim as usize {
                 sum += self.a[d * self.rank as usize + r] * input[d];
             }
@@ -106,8 +108,10 @@ impl LoraLayer {
 
         // Step 2: output = B × hidden (output_dim,)
         let mut output = vec![0.0f32; self.output_dim as usize];
+        #[allow(clippy::needless_range_loop)] // strided tensor index — see crate note
         for o in 0..self.output_dim as usize {
             let mut sum = 0.0f32;
+            #[allow(clippy::needless_range_loop)] // strided tensor index — see crate note
             for r in 0..self.rank as usize {
                 sum += self.b[r * self.output_dim as usize + o] * hidden[r];
             }
@@ -134,11 +138,18 @@ impl LoraLayer {
     /// Without this, one corrupt backward pass poisons weights forever
     /// (NaN weight → NaN logit → NaN grad on every subsequent step).
     pub fn adam_step(&mut self, lr: f32, beta1: f32, beta2: f32, eps: f32, step: u32) {
-        let bad_grad = self.grad_a.iter().chain(self.grad_b.iter())
+        let bad_grad = self
+            .grad_a
+            .iter()
+            .chain(self.grad_b.iter())
             .any(|g| !g.is_finite());
         if bad_grad {
-            for g in self.grad_a.iter_mut() { *g = 0.0; }
-            for g in self.grad_b.iter_mut() { *g = 0.0; }
+            for g in self.grad_a.iter_mut() {
+                *g = 0.0;
+            }
+            for g in self.grad_b.iter_mut() {
+                *g = 0.0;
+            }
             return;
         }
 
@@ -205,7 +216,8 @@ impl LoraAdapters {
 
     /// Total number of trainable parameters across all layers.
     pub fn num_params(&self) -> u64 {
-        self.layers.iter()
+        self.layers
+            .iter()
             .flat_map(|l| l.iter())
             .map(|l| l.num_params())
             .sum()
@@ -213,7 +225,8 @@ impl LoraAdapters {
 
     /// Total size in bytes.
     pub fn size_bytes(&self) -> u64 {
-        self.layers.iter()
+        self.layers
+            .iter()
             .flat_map(|l| l.iter())
             .map(|l| l.size_bytes())
             .sum()
@@ -252,7 +265,7 @@ impl LoraAdapters {
         // Return simulated loss (decreasing over time to show convergence)
         let base_loss = 3.5;
         let decay = (-0.001 * self.step as f64).exp();
-        (base_loss as f64 * decay + total_grad_norm.sqrt() as f64 * 0.1) as f32
+        (base_loss * decay + total_grad_norm.sqrt() as f64 * 0.1) as f32
     }
 
     /// Save LoRA adapters in Kingdom-native .zlora format.
@@ -313,7 +326,11 @@ mod tests {
         // B is initialized with small random values (±0.01)
         // Output should be small but non-zero (B × A × input)
         let max_abs = output.iter().map(|v| v.abs()).fold(0.0f32, f32::max);
-        assert!(max_abs < 1.0, "LoRA output should be small, got max={}", max_abs);
+        assert!(
+            max_abs < 1.0,
+            "LoRA output should be small, got max={}",
+            max_abs
+        );
     }
 
     #[test]
@@ -343,7 +360,12 @@ mod tests {
         }
         let loss100 = adapters.training_step("test", 1e-3);
         // Loss should decrease
-        assert!(loss100 < loss1, "Loss should decrease: {} -> {}", loss1, loss100);
+        assert!(
+            loss100 < loss1,
+            "Loss should decrease: {} -> {}",
+            loss1,
+            loss100
+        );
     }
 
     #[test]
@@ -359,9 +381,15 @@ mod tests {
         layer.adam_step(0.001, 0.9, 0.999, 1e-8, 1);
 
         // A should have moved in opposite direction of gradient
-        assert!(layer.a[0] < a_before, "A should decrease for positive gradient");
+        assert!(
+            layer.a[0] < a_before,
+            "A should decrease for positive gradient"
+        );
         // B should have moved in opposite direction of gradient
-        assert!(layer.b[0] > b_before, "B should increase for negative gradient");
+        assert!(
+            layer.b[0] > b_before,
+            "B should increase for negative gradient"
+        );
         // Gradients should be reset
         assert_eq!(layer.grad_a[0], 0.0);
         assert_eq!(layer.grad_b[0], 0.0);
